Add files via upload
This code submission addresses the testing of chip states by generating specific circuits for analysis. The code allows for the generation of a set of circuits to test the operational state of a chip. The circuits generated encompass two types: (1)GHZ-state generation circuits for any given $n$ number of qubits (a total of $n$ circuits), and (2)W-state generation circuits (both normal and inverted order, with up to approximately $4n$ CZ gates) for a comprehensive testing suite.  This suite of $n+2$ circuits offers a straightforward visual insight into the current operational state of the chip.